style: move all trajectory calculation into nuke util layer
